changes to allow token renewal
