+++ project/settings.py
@@ -336,7 +336,6 @@ REST_FRAMEWORK = {
     'DEFAULT_FILTER_BACKENDS': ['django_filters.rest_framework.DjangoFilterBackend'],
     'DEFAULT_AUTHENTICATION_CLASSES': [
         'project.npda.api.authentication_class.PDUScopedOAuth2Authentication',  # Custom authentication class for API
-        'rest_framework.authentication.SessionAuthentication',        # Keep for web app
     ],
     'DEFAULT_PERMISSION_CLASSES': [
        'rest_framework.permissions.IsAuthenticated',
@@ -353,16 +352,43 @@ OAUTH2_PROVIDER = {
         'patient:write': 'Write patient and visit data',
         'admin:cross-pdu': 'Admin access across all PDUs',  # Add this
     },
-    'ACCESS_TOKEN_EXPIRE_SECONDS': 86400,
-    'REFRESH_TOKEN_EXPIRE_SECONDS': 86400 * 7,
-    'REFRESH_TOKEN_GRACE_PERIOD_SECONDS': 120,
-    'ROTATE_REFRESH_TOKEN': True,
+    'ACCESS_TOKEN_EXPIRE_SECONDS': 86400, # 24 hours
+    'ALWAYS_RETURN_BASIC_ERRORS': True,  # Always return basic errors for security
+     'ALLOWED_GRANT_TYPES': ('authorization_code', 'client_credentials'),
+     'OAUTH2_PROVIDER_GRANT_MODEL': 'oauth2_provider.models.Grant',
 }
 # Global settings for Django OAuth Toolkit (using default models)
 OAUTH2_PROVIDER_ACCESS_TOKEN_MODEL = 'oauth2_provider.AccessToken'
 OAUTH2_PROVIDER_ID_TOKEN_MODEL = 'oauth2_provider.IDToken'
 OAUTH2_PROVIDER_APPLICATION_MODEL = 'oauth2_provider.Application'
 OAUTH2_PROVIDER_REFRESH_TOKEN_MODEL = 'oauth2_provider.RefreshToken'
+
+# OAUTH2_PROVIDER.update({
+#     'OAUTH2_BACKEND_CLASS': 'oauth2_provider.oauth2_backends.JSONOAuthLibCore',
+#     'RESOURCE_SERVER_TOKEN_CACHING_SECONDS': 3600,
+#     'CLIENT_SECRET_GENERATOR_LENGTH': 128,
+#     'CLIENT_SECRET_GENERATOR_CALLABLE': 'oauth2_provider.generators.generate_client_secret',
+#     'CLIENT_ID_GENERATOR_CLASS': 'oauth2_provider.generators.ClientIdGenerator',
+#     # Add this line to disable client secret hashing:
+#     'CLIENT_SECRET_VALIDATOR_CLASS': 'oauth2_provider.oauth2_validators.ClientSecretValidator',
+# })
+
+# Add to settings.py
+LOGGING = {
+    'version': 1,
+    'disable_existing_loggers': False,
+    'handlers': {
+        'console': {
+            'class': 'logging.StreamHandler',
+        },
+    },
+    'loggers': {
+        'oauth2_provider': {
+            'handlers': ['console'],
+            'level': 'DEBUG',
+        },
+    },
+}
 
 SECURE_SSL_REDIRECT = True
